Adding middleware flag to config.
