+++ config/feature-toggle.php
@@ -57,12 +57,14 @@ return [
      | Options to enable/disable behavior.
      |
      | [
+     |      'registerMiddleware' => true/false,
      |      'useMigrations' => true/false,
      | ]
      |
      */
 
     'options' => [
+        'registerMiddleware' => true,
         'useMigrations' => false,
     ],
 ];
