+++ src/ServiceProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace FeatureToggle;
 
 use FeatureToggle\Contracts\Api as ApiContract;
+use FeatureToggle\Middleware\FeatureToggle;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\ServiceProvider as SupportServiceProvider;
 
@@ -99,6 +100,8 @@ class ServiceProvider extends SupportServiceProvider
      */
     public function boot(): void
     {
+        $this->registerBladeDirective();
+        $this->registerMiddleware();
         $this->registerMigrations();
         $this->registerPublishing();
     }
@@ -113,6 +116,18 @@ class ServiceProvider extends SupportServiceProvider
         Blade::if('featureToggle', function (string $name, bool $status = true) {
             return feature_toggle($name, $status);
         });
+    }
+
+    /**
+     * Register the package migrations.
+     *
+     * @return void
+     */
+    protected function registerMiddleware(): void
+    {
+        if ($this->app['feature-toggle.api']->isMiddlewareEnabled()) {
+            $this->app['router']->aliasMiddleware('featureToggle', FeatureToggle::class);
+        }
     }
 
     /**
