+++ src/Traits/HasStaticOptions.php
@@ -14,6 +14,54 @@ trait HasStaticOptions
     protected static $options = [];
 
     /**
+     * @return bool
+     */
+    public function isMiddlewareEnabled(): bool
+    {
+        return filter_var($this->getOption('registerMiddleware', false), FILTER_VALIDATE_BOOLEAN);
+    }
+
+    /**
+     * @return bool
+     */
+    public function isMigrationsEnabled(): bool
+    {
+        return filter_var($this->getOption('useMigrations', false), FILTER_VALIDATE_BOOLEAN);
+    }
+
+    /**
+     * @return void
+     */
+    public static function useMigrations(): void
+    {
+        static::setOption('useMigrations', true);
+    }
+
+    /**
+     * @return void
+     */
+    public static function ignoreMigrations(): void
+    {
+        static::setOption('useMigrations', false);
+    }
+
+    /**
+     * @return void
+     */
+    public static function registerMiddleware(): void
+    {
+        static::setOption('registerMiddleware', true);
+    }
+
+    /**
+     * @return void
+     */
+    public static function ignoreMiddleware(): void
+    {
+        static::setOption('registerMiddleware', false);
+    }
+
+    /**
      * @param  string  $name
      * @param  null|mixed  $default
      * @return mixed
